+++ lib/onlyoffice_digitalocean_wrapper/digitalocean_wrapper.rb
@@ -25,13 +25,6 @@ module OnlyofficeDigitaloceanWrapper
       raise ArgumentError, 'DigitalOceanWrapper: Your Access Token is Incorrect' unless correct_access_token?
     end
 
-    def current_kernel(droplet_name)
-      droplet = droplet_by_name(droplet_name)
-      kernel_name = droplet.kernel.name
-      OnlyofficeLoggerHelper.log("get_droplet_kernel_by_name(#{droplet_name}): #{kernel_name}")
-      kernel_name
-    end
-
     def wait_until_droplet_have_status(droplet_name, status = 'active', params = {})
       timeout = params.fetch(:timeout, 300)
       counter = 0
@@ -44,13 +37,6 @@ module OnlyofficeDigitaloceanWrapper
       raise DropletOperationTimeout, "#{droplet_name} was not #{status} for #{timeout}s" if counter >= timeout
 
       get_droplet_status_by_name(droplet_name)
-    end
-
-    def kernels_of_droplet(droplet_name)
-      droplet_id = get_droplet_id_by_name(droplet_name)
-      kernels = client.droplets.kernels(id: droplet_id).to_a
-      OnlyofficeLoggerHelper.log("Got kernels_of_droplet(#{droplet_name})")
-      kernels
     end
 
     def restore_image_by_name(image_name = 'nct-at-stable',